sibling of 0191f16ebca1250f8accbb939621d47af4e70c58

+++ crates/assistant_context_editor/src/language_model_selector.rs
@@ -664,7 +664,7 @@ mod tests {
             format!("{}/{}", self.provider_id.0, self.name.0)
         }
 
-        fn max_token_count(&self) -> u64 {
+        fn max_token_count(&self) -> usize {
             1000
         }
 
@@ -672,7 +672,7 @@ mod tests {
             &self,
             _: LanguageModelRequest,
             _: &App,
-        ) -> BoxFuture<'static, http_client::Result<u64>> {
+        ) -> BoxFuture<'static, http_client::Result<usize>> {
             unimplemented!()
         }
